+++ lib/jump_in.rb
@@ -2,6 +2,8 @@ require "jump_in/version"
 require 'jump_in/authentication'
 require 'jump_in/password_reset'
 require 'jump_in/tokenator'
+require 'jump_in/authentication/session'
+require 'jump_in/authentication/cookies'
 
 module JumpIn
